requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/rense/django-jet#egg=django-jet
-e git+http://github.com/russ-/pychallonge#egg=pychallonge
bcrypt==3.1.4
certifi==2018.8.24
cffi==1.11.5              # via bcrypt
chardet==3.0.4            # via requests
django-autocomplete-light==3.3.0
django-ckeditor==5.6.1
django-debug-toolbar==1.9.1
django-dotenv==1.4.2
django-js-asset==1.1.0    # via django-ckeditor
django-recaptcha2==1.3.0
django==2.1
djangorestframework==3.8.2
feedparser==5.2.1
gunicorn==19.9.0
idna==2.7                 # via requests
iso8601==0.1.11
pillow==5.2.0
psycopg2==2.7.5
pycparser==2.18           # via cffi
python-http-client==3.1.0  # via sendgrid
pytz==2018.5              # via django
raven==6.9.0
reportlab==3.5.6
requests==2.19.1
sendgrid-django==4.2.0
sendgrid==3.6.5           # via sendgrid-django
six==1.11.0               # via bcrypt
sorl-thumbnail==12.4.1
sqlparse==0.2.4           # via django-debug-toolbar
urllib3==1.23             # via requests
